Enqueue digest runs in groups
This allows some workers to start processing digests because some
will already be committed to the database.

Currently it takes about 10 minutes for this process to import
and queue up all the jobs, so if we can start queuing up jobs
before that we can process all the digests quicker.