Remove QueueLatencyHealthcheck
We're having trouble setting sensible limits for this as the expected
latency for the digest queues can be up to 1 hour at the moment which is
way too high for immediate queues.

Removing it for now to be reinstated as two checks, one for digest
queues and one for immediate that can have limits set independently.