DeliveryAttempt status is set by provider
This changes the provider classes to return the status of the delivery
attempt they just made. This is since we expect different values based
on providers (psuedo should only return :delivered, whereas notify
should return :sending or a :technical_failure)

The inspiration for this was discovering that we set a permanent status
of :sending for emails sent with psuedo provider.

This allows us to drop the need for the ProviderError exception and
handle this directly in the notify provider class.

This does introduce an annoyingly level of coupling between
DeliveryAttempt and Provider classes whereby Provider classes need to
know valid status values as their return values. I'm not sure right now
of a better route though. Options seem to be: have providers create the
delivery attempts (deeper coupling) or have providers return an object
with generic details (which seems to basically be purpose of delivery
attempt).