Refactor status update service for error handling
This refactor of StatusUpdateService was initially intended to make it
preload the delivery attempt and email models so they could be used as
part of the Email finalise process.

Along the way though I caught a few conditions that could result in
errors or affecting data integrity so this switched track somewhat to
take them on.

This changes what happens when we receive a status that we don't
understand from Notify from us returning a 500 to us instead returning a
422. We also log this with Sentry.

This addresses a different pain point we have too where it is in theory
possible that a delivery attempt status could be updated multiple times
by Notify, whereas we only expect to get the callback once. This has now
been configured to return a 409 if a delivery attempt already has a
status that is not "sending".