Ensure order of emails and subscription contents match
Previously we were inserting the emails based on the order of the
subscribers, and matching them back with subscription contents
based on the order of subscription contents (which was undefined).
Most of the time this seemed to work, by luck, but we were seeing
it fail a few times in tests (and likely in production too).

Instead we now make sure that we iterate over the data structure
in the same say both times by introducing a new method called
'map_subscriber_content_change_id_in_order' which is used by both
the email inserting and matching back with subscription contents.