EmailArchiveWorker class
This class iterates through all emails that are archivable and archives
them into the EmailArchive table. It runs in batches of 1000 but there
is no limit, so if this hasn't been run in a while it could be a rather
painful task.

It serves the puprose of using data from EmailArchiveQuery to populate
the EmailArchive table and then marks items in the Email table as
archived which should allow them to be deleted at a later date.