Configure EmailArchiveWorker with sidekiq
This schedules this worker to run every 1 hours to mark emails as
archived.

It could transpire that this is better to run at a quiet time but I
decided to try do this throughout the day so it could hopefully be more
of a little and often task that's easier to swallow and reveals issues
earlier than one that may be harder to notice in the night.

On my local machine this is the performance: "Archived 323912 emails in
178.22 seconds". The memory usage seemed to very reasonable.