EmailDeletionWorker
This worker is used to delete email data that has reached the criteria
where it was deleteable.

It is scheduled to run every hour and like EmailArchiveWorker it could
be argued that this is better to run at a quiet time. In my tests though
it does seem to run better regularly than large runs.

On my local machine it did: "Deleted 323912 emails in 219.61 seconds".

I considered, but decided against, running deletes on the various tables
associated with an email first as it felt like this was me trying to out
optimise Postgres - it could well be that that is more efficient.