Add sent_at and completed_at to DeliveryAttempt
Email-alert-api now records when the email was delivered by notify 'sent_at'
and when a status update occurs 'completed_at', e.g from temp-failure to
permanent failure. This information is recorded in the DeliveryAttempt
model. Also on the Email model, the 'finished_sending_at' column is now
populated with the 'sent_at' value, instead of the 'updated_at'.