Isolate faking Sidekiq with an around block
Our default state for tests set in `spec/rails_helper.rb` is for Sidekiq
to be `inline`. The notification handler service test relies on setting
fake` to be true. Use an around block so we properly isolate this and
don't leak the state through to the rest of the tests.