Fix broken SubscriptionContentWorker test
Now that Sidekiq is working as intended and running `inline`, we have a
small testing anomaly:

We stub ImmediateEmailBuilder's interaction which means an Email is not
created in the database. When DeliveryRequestWorker tries to find
that email, it fails.

Fix this - in this instance we don't really care about
DeliveryRequestWorker so we can stub it to return true.