Properly isolate the temporary failure state in testing
When an email has a temporary failure, we set it to retry delivery in 15
minutes time. We want to  test that during the temporary failure state, we
do not set a `finished_sending_at` time on the Email. We have set up
`Sidekiq::Testing.inline!` in our `spec/rails_helper.rb` which means any
Sidekiq jobs are immediately fired. The side effect of this is when
testing a temporary failure, we end up immediately sending a new
delivery attempt which succeeds and sets the `finished_sending_at` field
on the Email before we can assert that it is `nil`. To avoid this
behaviour, we wrap this test in `Sidekiq::Testing.fake!` which pushes
jobs on to an array rather than immediately firing them.