Create enum fields on emails for status and failure_reason
This is part of adding in the feature to stop retrying emails after a
certain number of attempts. To store the difference between a delivery
attempt that is a final state and a temporary state we either need to
store accumulative state on a DeliveryAttempt or utilise Email for a
view of all DeliveryAttempts.

This moves this to be stored on Email so that we don't have to store
DeliveryAttempts that have failed for different reasons with different
states.