Add enums for email state and failure reason
This is to allow us to de-couple the difference between an email status
on a delivery attempt (which can be cumulative) with that of the email
as a whole.

It uses a pending status rather than a sending status as an email isn't
initially sending and I didn't think we needed to distinguish between
that and sending (but maybe we'll decide that's useful later).