Allow failed timestamps to update Email finished_sending_at
By storing delivery_attempt.sent_at on Email we were at risk of setting
this to be nil on Emails that failed to send. By setting this to sent_at
or completed_at we have a much lower risk of this.

As we're not actually validating that these always send through a
timestamp we're still at a small risk that if Notify don't send through
a timestamp we'll end up with inconsistent data. Some of the follow up
PR's I'm working on for Email status should alleviate this risk.