Don't create an index on subscriber_id
This modifies an existing migration, but it's not yet deployed
and we need this change this migration so it doesn't lock the
email table.

Unfortunately, when I wrote the migration, I was looking at the
documentation for an older version of Rails where the default was
to not create an index, but this has changed in Rails 5 to create
an index by default. I didn't notice because it ran fast locally
and on integration, but on staging the deploy is taking a very
long time and locking the table.