Add basic AuthTokenGeneratorService
This is a simple service that takes a subscriber and will return a JWT
token that is valid for an unlimited period of time. It is intended for
Email Alert Frontend, but could be applicable to other situations.

Likely we'll want to update this to have an expiry time and we may want
additional fields within it.