Don't send digests without content changes
This is a safeguard to prevent us from sending out digests that don't
have any content changes.

Subscribers should not be queued for digests if they don't have any
content changes but this is not atomic and accidents can occur (as per
the ~6000 blank emails we sent out over the previous weekend) so this
can perform a no-op on a queued digest run subscriber that shouldn't
actually receive emails.

This commit also does a fair bit of refactoring of this worker class to
make it more procedural and reduce indirection. It does this by
preferring arguments for variables rather than using instance variables.