Add basic validation layer for auth token generation
This validates the input and relies on Rails to wrap this into a 422
response.

The Rails response handler will only show the status and name and not
the actual messages, this seems fine right now as this will be logged in
Kibana. However to be a good API this should really return the error
details.