Relax required sent_at param for StatusUpdateService
When Notify send us status updates, the `sent_at` paramter can be `nil`.
Currently we return a 400 when `sent_at` is `nil` and do not update
the database.

Allow `sent_at` to be `nil` or absent in StatusUpdateService so we
accept these status updates without error and write them to the db.