Don't create accounts for users who don't exist
This changes the behaviour of the /subscribers/auth-token endpoint to
return a 404 when it tries to look up a subscriber that is not known in
the system. This previously created a record for that subscriber.

The reason for this is to store a minimal amount of personally
idenitifiable information on a user of this system.