SubscriptionShowPresenter class
This is to replace the increasing complex logic that is going into the
as_json method on Subscription to serialize it for rendering as part of
HTTP responses.

It also included an ended boolean value that can be used to work out if
a subscription has ended for dealing with unsubscribing.