Switch Time.now methods to be Time.zone.now for DigestRun
This means that all the Time objects we create are in the zone that
Rails is configured for (Europe/London). When these are created without
Time.zone these default to the servers time (which is UTC) and this
causes our validation issues.

Ideally we should be using Time.zone.now everywhere rather than Time.now - I
think there is even a Rubocop cop that checks for this.