Anonymise email addresses in subjects/bodys
A problem that we currently have is that our sanitisation script doesn't
replace instances of email addresses that are within emails themselves
which can render some of our other anonymising work superfluous.

This uses the postgres replace function to change references to the
subscribers email address in subject and body parts of the email.

I'm hoping to try determine the effects this has on performance but this
is quite difficult as performance without this is currently so poor.