Create worker to mark digest runs as complete
After every digest email is sent, the worker checks if it was the last
email to be sent for that digest run. If it is, it marks the run as
complete. The check required is expensive, between 2 to 4 seconds in
the database while under load. We’ve been unable to make postgres use
an index to speed up the operation.

This commit adds a new worker that finds all runs that haven’t yet been
marked as completed, checks if all the subscribers are marked as
completed then uses the most recent completion time as the run’s
completion time. Running this every minute will still give us accurate
statistics and will greatly reduce the load on Postgres.