Remove uniq validation for title in SubscriberList
Signing up for email alerts for certain taxon's throws an error due to the uniq
constraint on title in the SubscriberList model. This validation was
needed when Email was using GovDelivery, but now serves no purpose.

The removal of this validation will allow users to sign up for email
alerts for taxons which share the same title as other legacy taxonomies
instead of returning an error.