Change the batching approach in EmailArchiveWorker
Don't use in_batches, as this sorts by the id (which is a UUID) which
adds unnecessary overhead. As the emails are archived, the
EmailArchiveQuery will ignore them when fetching the next batch, so
the approach can be simplified to just fetch batches in a loop, and
then terminate after the final batch is processed.