Increase subscription content health check thresholds
We had tried to do this in the past but it wasn't increased enough.

At around 9:30 every morning we regularly see this health check go
critical because of a large number of scheduled publishings go out
and it takes a bit of time to process all the emails. Although the
alert is going critical, there's no problem for the 2nd line team
to deal with so it's confusing that the alert appears.