Add missing policies links to policy subscriber lists
Recent Policies seem to be missing the relevant links. I'm not sure
what effect this is having, or why they are really missing, but as
we're working on retiring Policies, I've just been looking at fixing
this data, so that these policies can be retired.

This migration adds the missing links.