Track unpublish emails
Set the ended_email_id on the subscription when a unpublish message
comes in.

This involved quite a bit of refactoring (which Jos and I worked on
together). Each subscriber list is processed in turn for simplicity,
and the updating of the subscriptions is now part of the
UnpublishHandlerService as it has the email_id's.

The deactivation of subscribers is done through a worker, to avoid
doing to many things at once.