Support ending subscriptions in bulk
This is a bulk approach, intended to compliment the immediate approach
of ending subscriptions when the thing a user subscribes to is
unpublished.

This approach will take a set of things, and a URL to point the user
at if they want to continue receiving updates, and will send one email
to each subscriber detailing all the subscriptions that they'll no
longer receive updates for (because they have ended).

The intent is to run this prior to unpublishing large numbers of
things people subscribe to (at first, Policy pages). For users who
have quite a few (e.g. > 10) subscriptions that will be ending, this
will hopefully be a better experience than receiving lots of
individual emails.