Add 'any' option to links and tags in subscriber lists.
For multivalued links such as taxonomy_tree we want to introduce
functionality to subscribe to changes using an 'and' operator, i.e.
subscribe to changes in documents that are tagged to the 'Brexit'
_and_ 'Transport' taxons. Currently the email-alert-api uses an
implicit 'or' operator so documents would be returned tagger to
the 'Brexit' _or_ 'Transport' taxons.

This commit makes the 'or' operator explicit by adding an 'any' key to
the links and tags hashes in the SubscriberList models.