Adds 'all' operator to all tags and links in subscriber lists.
We want to subscribe to changes in multiple values of the same link/tag,
i.e. subscribe to changes in documents that are tagged to the 'Transport' i
taxon _and__ the 'Brexit' taxon.

This commit implements this by adding an 'all' key to the links/tags
hash in the SubscriberList model