Allow rejection-type filtering of content_purpose_supergroup
This lets one set reject_content_purpose_supergroup as a parameter
when creating and looking up subscriber lists.

This parameter, if set, will enable an email subscription to not
receive emails when documents in certain supergroups are changed.

The main use case for this is the other supergroup, which is the
default.