Count subscriptions for each sector
Adds a rake task to count the number of subscriptions to the business
readiness finder for each sector/business area.

e.g.

```
bundle exec rake export:csv_from_sectors_in_business_readiness
```

Example output:

```
title,count
"Accommodation, restaurants and catering services",63
Aerospace,47
Agriculture,60
Air transport (aviation),54
.
.
.
```

When the results are generated, each subscription may be counted more
than once. This is because email-alert-api doesn't store subscriptions
against the slug of the page. It instead slugifies the title of the
page. That means that the slug in email-alert-api may not match the slug
of the page.
In the case of finder pages, email-alert-api generates an "and" query
and concatenates all of the facets to the end of the slug. Thats means
email subscriptions to the busniness readiness finder could have a slug
like:

"find-eu-exit-guidance-for-your-business-with-sector-business-area-of-accommodation-restaurants-and-catering-services-and-retail-business-activity-of-i-sell-products-or-goods-in-the-uk-and-employ-eu-citizens-of-no"

As the slug contains two sectors/business areas it will be counted twice
in the script. Once for
“accommodation-restaurants-and-catering-services” and once for “retail”.

All of the slugs for the business readiness finder have the same prefix.
To allow the tests to be generic and prevent failures if the the slug of
the finder changes, the slug prefix has been made publically available.

Ideally we'd want to remove the querying by slug and instead query by
the sector facet. The selected facets are stored in the subscription as:

`tags: {"sector_business_area"=>{"any"=>[“accommodation-restaurants-and-catering-services”, "retail"]}`

As they are JSON fields, the tags are not simple to query, but this
would be a good enhancement to make when there is more time.