Add rake task to find subscription lists by title
This rake task will search for `SubscriberLists` using a partial
or whole title match, and will return the `title` and `slug` for
each one.

To be used to assist with the rake task
`manage:update_subscriber_list[slug,new_title,new_slug]` as part
of jobs such as renaming a county.