Generate report on successful delivery attempts
We want to monitor how long it takes for an email to be delivered
successfully by comparing the `created_at` and `updated_at` fields.

This script gets all successful delivery attempts within the given
dates and compares the two fields, writing the time difference to a
csv. It also outputs the average of these values to the terminal.

It's worth noting that emails and their delivery attempts are archived
after 24 hours of an email reaching it's final state (sent/failed) so
this script can only be run for the past day.