Add GovukDocumentTypes to content_change tags
This will add attributes like content_purpose_subgroup to
a content_change's tags. This will make it possible to
subscribe to groups of documents defined in the GovukDocumentTypes
gem.

A previous way to make it possible to subscribe to these attributes was to add
a new column content_purpose_supergroup to the subscriber_lists
relation and then filter using that field. This can be removed.