Move subscriptions from legacy subscriber lists
We will soon remove the 'content_purpose_supergroup' field in
the SubscriberLists. The content_purpose_supergroup is now
embedded in the 'tags' field.

This commit introduces rake tasks to get rid of any subscriber
lists with a value set in content_purpose_supergroup field.

This process is split into two tasks. The first moves all subscriptions
from the 'legacy' subscriber_lists to the new lists. The second removes
the now unused subscriber lists and moves all matched_content_changes
to their new counterparts. This is done so we avoid most
synchronisation issues that may arise