Implements FindLatestMatchingSubscription query & tests
This writes a query which fetches the most recent subscription that
matches the given subscription, i.e. has the same subscriber, subscription
list and frequency. In most cases this will simply return the same
subscription, but it's intended to be used to detect unsubscribe requests
from old subscription emails that the user may have clicked on, as this is
the same subscription conceptually.

Assumptions:

- When a user changes their subscription through frequency change
- And when a user changes through unsubscribing and resubscribing
- A new subscription is created every time, with a more recent 'created_at'
- In other words, it's not possible for an older record that has become
inactive to become active again. It will always be a new subscription