Removes frequency from FindLatestMatchingSubscription query
This query originally matched on subscriber ID, subscription list and frequency, such that
if a user subscribes to something (creating Subscription A), then unsubscribes and later
resubscribes (creating Subscription B), then we can use FindLatestMatchingSubscription to
query Subscription A and return Subscription B (which matches).

This was incorrectly implemented to make frequency part of the match, but we actually need
to respond differently if the latest subscription is a different frequency, which makes this
query less useful. In order to implement all of the business logic required in
https://github.com/alphagov/email-alert-frontend/pull/474, we need to return the latest
subscription that matches the given subscription list and subscriber ID, irrespective of its
frequency.