Revert "Adds or joined facet subscriber list"
But retains the database migration, update to schema and
refactor of tests

We implemented a OrJoinedFacetSubscriberList model to enable
'OR' logic queries between facets in the business readiness
finder.  However we ended up implementing this using the
existing 'any or all' logic across facet_values leaving the
OrJoinedFacetSubscriberList redundant.  Therefore this code
can be removed.