Add migration to remove type column from subscriber_lists
The type column was added to allow us to use single table inheritance
for subscriber lists (to create a OrJoinedFacetSubscriberList) in order
to allow 'OR' logic between facets.

This logic was never used, so the OrJoinedFacetSubscriberList model
has been removed.  This migration removes the database column used
to identify this model.