Quicken SubscriptionContent healthcheck
Previously running this healthcheck involved querying two large tables
twice: SubscriptionContent and Subscription. Since we expect the steps
taken to address the alert to be identical for both a warning and a
critical, this reduces the runtime of the check by just having one
(critical) severity level, but using the parameters for the (deleted)
warning severity level so that the alert will still be timely.