Deleting a SubscriberList will delete associated subscriptions
We need to delete invalid subscriber lists, so this makes
that possible and ensures referential integrity.

It also deletes subscriptions contents. As these are invalid
subscriber lists it's not expected that there will be any
but this is a just-in-case and ensures we don't have any
orphaned subscription_contents.

In a subsequent commit I'm going to add a rake task to
delete invalid subscriber lists.