Perform one query instead of two for subscription contents healthcheck
This query can sometimes take a long time (5 seconds or more). We're not
entirely sure why this is, the query plans all say that the necessary
indexes are being used.

Currently we end up doing the same query twice, once for criticals and
once for warnings. When they take 5 seconds each, this can often lead to
the healthcheck endpoint timing out.

To avoid the healthcheck taking too long, we can perform the query
once (for the warnings) and then count the number of those warnings
that would also be classes as criticals.