Update subscriber lists from `EU Exit` to `Brexit`
We are changing all instances of `EU Exit` to `Brexit` across
GOV.UK.

This rake task finds all subscriber lists that come from the business
readiness finder that contain `EU Exit` in the title and slug, and is
replaced with `Brexit`. The changes will be reflected in the affected
email templates.