Add additional permitted tags
This is the result of an audit of the applications that can
send tags downstream to email alert api. It looks like many
of the permitted tags haven't been used yet.

See https://gov-uk.atlassian.net/wiki/spaces/TECH/pages/17989732/Email+alert+technical+overview
for more details of the architecture.