Create a message model
This model is intended for ad-hoc messages to subscriber lists to send
them a particular message. This could be used as a standard mailing list
message or as a means to notify them to a change to a particular list.

They are associated with links and tags so they can be matched up via
the same approach as other emails. I'm not currently sure if that is the
best approach or not.