Move SubscriptionContentWorker to ProcessContentChangeWorker
This moves the code from SubscriptionContentWorker to
ProcessContentChangeWorker so that a second worker can be created to
process messages which will also generate subscription contents.

This also serves as a refactoring of the code to reduce the verbosity.
The import logic is moved and simplified in the model (it can be
simplified further once we upgrade past postgres 9.4), it also uses active
record scopes rather than specific query class
ContentChangeImmediateSubscriptionQuery.