Create query to look digest subscription content
This is to act as a replacement for SubscriptionContentChangeQuery with
the view that this can pull data out for both content changes and
messages.

This will also allow the removal of the ContentChangeDeduplicator
service which seems to have had limited value since
https://github.com/alphagov/email-alert-api/pull/588 was merged in.

A somewhat opinionated change that was made as part of this was to order
results by created_at value rather than public_updated_at. This has been
done since this is a date that can be compared with messages. This also
better represents the time these alerts were created, which is probably
beneficial given there is nothing to stop the public_updated_at value
falling outside the digest date. For documents without a manipulated
public_updated_at value this change should present no difference.