Allow slug to be passed in to create subscriber list
This commit makes passing in a slug when creating a subscriber list
possible, this is done so the slug can be unique to the title.

We need to do this as we can't control the length of the title and so
are setting a generic title for the dynamic lists subscriptions so the
user doesn't receive a big long string that makes no sense.

Trello:
https://trello.com/c/Wsm83fy7/56-build-email-signup-workflow