Query to look up subscriber lists by criteria
Given a set of criteria rules look up a subscriber list that matches
them.

This lookup is performed by using recursion to build an ActiveRecord
query, this allows only looking up SubscriberLists that exactly match
the criteria and not having to do any determining through ruby. However
this is at the detriment of the ease to understand exactly how the query
is built as there is a lot of scope conditions in play.