Move matching messages to be done at a worker
Since we have to process the message asynchronously we may as well do
this as part of the asynchronous work and not risk delaying the request
if the query is slow.

It's simple for us to do this for messages as they're not in use at the
moment whereas this would be a more complicated change to deploy for a
content change as this would cause problems for any that already being
processed.

This also removes the somewhat superfluous keyword argument on the
MatchedMessageGenerationService as it added unnecessary verbosity.