Remove unused fields from message
Since we don't have a use case for messages without criteria_rules all
of these fields currently don't serve a purpose.

If the needs for Message are to broaden in the future then we can always
add these back, but there doesn't seem much value in them sitting around
un-used.