Rewrite FindExactMatch query to use tags/links digests
This query should be more performant now. Rather than requesting
all lists with the same keys as requested, and looping through them
in ruby, we lookup a list on the digest of the JSON.

I ran a benchmark to demonstrate the performance improvement.

I created 25k subscriber lists locally, and ran both master and this
branch, and fired requests at it using `ab` (Apache Benchmarking tool).

ab -t 10 -r -c 5 http://email-alert-api.dev.gov.uk?/subscriber-lists?tags...

The test was quite biased towards demonstrating the best case for this test.
I created 25k subscriber lists with the same tag keys, but different values.
We likely have a situation similar to this in production, as we have many
more subscriber lists, with subscriptions mostly on the same sets of tags.

Results of benchmarking:

Master:

18 requests completed.

Percentage of the requests served within a certain time (ms)
  50%   2518
  66%   2549
  75%   2744
  80%   2873
  90%   3138
  95%   3161
  98%   3161
  99%   3161
 100%   3161 (longest request)

This branch:

1167 requests completed.

 Percentage of the requests served within a certain time (ms)
   50%     41
   66%     45
   75%     47
   80%     49
   90%     53
   95%     57
   98%     63
   99%     66
  100%    598 (longest request)