Try make JSON responses more consistent
The horse has already bolted on this where there is very little
consistency with responses and there are many that can't be changed
without BC breaks.

As a small step it sets all error responses to have a field of error and
allows a field of details for extra information. I don't expect this
will cause any problems for API users as I don't believe anything relies
on parsing these.