Try out list_group_id as a name for new SubscriberList field
This creates a new field list_group_id on the subscriber_lists table.
The field is to be intended to be used as a means to provide a common
grouping for a family of lists that are variations of the same thing.
It is intended to be used as a means to query groups of related lists,
but vary by criteria. Thus it could be used as a means to look up all
the users who a signed up to a particular finder but not specific
criteria within it.