Resolve consistency issues with subscriber list tests
This resolves an issue where we had a subscriber_lists_controller_spec
that had creating and showing, a get_subscriber_list_spec for showing
and a different create_subscriber_list_spec.

These tests are still a bit weird as we've got the controller tested in
scattered around files, but resolving this seems a larger piece of work
on consistency in this applications testing which doesn't seem something
that'd be fast to resolve.