Reduce length of existing slugs
This reduces all the slugs in the db to be up to 255 characters. It uses
the same logic approach as is used to create the slugs however it
doesn't attempt to put this code into a shared reusable place as this
migration is at a particular point in time and if the slug logic changes
then this migration no longer makes sense.