Remove limits on title and slug fields
On multiple occasions the limits for these fields have been increased
without any application of the limits in the local application. With the
change to reduce the limit of slug (given they should now be 255
characters max) I was faced with the choice of whether to update the
limit again or not.

Given it's inconsistent for us to apply limits at a PostgreSQL level I
feel these should be removed and if we need to apply limits this should
be done at an application level. This makes these fields consistent with
the other string fields in Email Alert API.