Order subscriptions
This is needed as we are adding when subscriptions have been created on
the subscription management page in email-alert-frontend and it makes
sense for the order to default to created_at descending.

The api has also been updated to allow an order parameter to be passed
for either the `created_at` field for a subscription or the `title` for
a subscriber list. The syntax for descending or ascending values is as
follows `-title` for descending and just `title` for ascending. This
makes the endpoint less coupled to email-alert-frontends code and more
flexible if this endpoint was to be used by something else.

This specific action seemes to be only used for the subscription
management flow, from a git search and a local grep.

Trello:
https://trello.com/c/FGYU4NlN/92-differentiate-between-your-get-ready-for-brexit-results-in-subscription-management