Run healthcheck inside a controller
This allows this request to be logged by the Rails logger which is
otherwise omitted when a rack response is returned for a route.

We have seen timeouts for the Healthchecks for Email Alert API so it'd
be useful to have logging information available to determine how long
these are regularly taking.