Add a partial index on the subscription_contents table
This should improve the performance of the
SubscribersForImmediateEmailQuery, as it looks through the
subscription_contents table for rows where the email_id is
NULL. Previously PostgreSQL would probably do a Bitmap Heap Scan over
the table, however with this new partial index it can just do an Index
Only Scan across the index.

The performance of this query is important, as it's run regularly
through the ImmediateEmailGenerationWorker.