Add a load testing task to create delivery attempts
We want to do some load testing of the public /status-updates endpoint
and for that we need a lot of delivery attempts. This task makes it easy
to create lots of fake delivery attempts and then writes the references
to a file so the endpoint can be used.

The reason we need to create many delivery attempts is that the endpoint
will refuse any status update on a delivery attempt which has already
been updated, so we need many delivery attempts available to set the
status on.