Add subscriber information report
Add a report that will take one or more email addresses and return a
JSON payload of subscription information. This will include information
about the subscriber, their subscriptions and the associated
subscriber lists.

The report simply pretty prints to the console. I considered saving to a
file, but the reports are usually fairly one-off and the extra hurdle of
getting onto a machine and grabbing the file seemed unnecesary when we
can just run the task in Jenkins and see the results.