Add class to immediate send emails for a content change
Add a class that will allow us to force send emails for a content
change.

This is useful when, for example, travel advice emails have not gone out
for some reason and we want to quickly try and send them.

Use the query to find unsent emails for a content change id, then for
each of these, queue a delivery request on the immediate high queue.