Generate emails in ProcessWorkers
In response to an incident in which a
Travel Advice email was stuck in a backlog
of ~1 million unprocessed SubscriptionContents.
The bottleneck was that there was only
one ImmediateEmailGenerationWorker and it
could not process SubscriptionContents
concurrently.

By moving the generation of emails to the
ProcessContentChangeAndGenerateEmailsWorker and
ProcessMessageAndGenerateEmailsWorker, it will now be able
to create emails concurrently, thus
removing the bottleneck. It will also make
the system conceptually simpler and make it easier
to see where emails are in the system as the
length of job time or number in the queue
will be more visible.

In the last year, the most content changes
created in a single day was 1581 and the mean
per day was 277 so although this will increase
the time these new workers will take to run,
it shouldn't cause huge backlogs to build up.

The existing ProcessContentChangeWorker
and ProcessMessageWorker and
ImmediateEmailGenerationWorkers are left in.
This is to make the deploying process simpler
and so existing jobs can continue to run and be
restarted if necessary. A few months after
this is merged, we can remove the old workers